+++ CHANGELOG.md
@@ -8,9 +8,11 @@
 ### Changed
 - a few common timezone abbreviations localized for German (MEZ, MESZ, OEZ, OESZ, WEZ, WESZ)
 - better timezone search keyword coverage
-- binary clocks now allow to turn seconds off, like seven-segment clock
+- graphical binary clocks now allow to turn seconds off, like seven-segment clock
+- Swatch beats now allow to show two decimals
 
 ### Fixed
+- Swatch beats were up to half a .beat off (because of unwanted rounding)
 - custom labels were reset to default labels on repeated edits
 
 ---